Refactored profile_routes.py to: Call resolve_user_ids() once at the start (line ~79) Reuse the user_ids list for all subsequent queries Inline the rental fetch logic to reuse user_ids instead of calling get_current_rental_internal() which calls resolve_user_ids() again Added better error logging with exc_info=True Rewrote useEffect dependency logic in profile/index.tsx Fixed monthly_rent type: float → int in rental_routes.py Optimized resolve_user_ids reuse to -1 Supabase query per request in profile_routes.py Fixed p2p_loans FK disambiguation in banking_routes.py Eliminated 3-query pattern → 2 queries with joins in education_routes.py

+++ my_flask_app/app/routes/rental_routes.py
@@ -183,7 +183,7 @@ def rent_property(current_user_id: str):
             'id': rental_id,
             'player_id': current_user_id,
             'property_id': str(data.property_id),
-            'monthly_rent': float(monthly_rent),
+            'monthly_rent': int(monthly_rent),
             'is_active': True,
             'rented_at': datetime.utcnow().isoformat()
         }).execute()
